+++ packages/example/src/gatsby-theme-carbon/components/Header.js
@@ -3,7 +3,7 @@ import Header from 'gatsby-theme-carbon/src/components/Header';
 
 const CustomSwitcher = () => (
   <Header>
-    gatsby theme <span style={{ fontWeight: 200 }}>&nbsp;Carbon</span>
+    <span>Gatsby theme</span>&nbsp;Carbon
   </Header>
 );
