+++ packages/gatsby-theme-carbon/index.js
@@ -1,14 +1,13 @@
 // Generic MDX Components
 export { AnchorLinks, AnchorLink } from './src/components/AnchorLinks';
 export { default as PageDescription } from './src/components/PageDescription';
-export { default as Video } from './src/components/Video';
+export { default as Video } from './src/components/Video/Video';
 export { default as DoDontExample } from './src/components/DoDontExample';
 export { Row, Column, Grid } from './src/components/Grid';
 export { default as Caption } from './src/components/Caption';
 export { default as ResourceCard } from './src/components/ResourceCard';
 export { default as ArticleCard } from './src/components/ArticleCard';
 export { default as Aside } from './src/components/Aside';
-// export { default as ImageComponent } from './src/components/ImageComponent'; // in progress
 
 // Homepage Template Components
 export { HomepageCallout, HomepageBanner } from './src/components/Homepage';
