+++ packages/gatsby-theme-carbon/src/components/PageHeader/page-header.scss
@@ -9,7 +9,6 @@
 }
 
 .page-header__title {
-  padding-left: 1rem;
-  margin-bottom: $layout-xs;
+  margin-bottom: $spacing-06;
   margin-top: auto;
 }
