+++ packages/gatsby-theme-carbon/src/components/Video/index.js
@@ -1,0 +1,3 @@
+import Video from './Video';
+
+export default Video;
