+++ packages/gatsby-theme-carbon/src/components/Video/video.scss
@@ -1,0 +1,3 @@
+.#{$prefix}--video {
+  margin-bottom: $spacing-08;
+}
